move BA and CI classes under a common Infantry abstract class, moving CI-specific methods to ConvInfantry
