+++ megamek/src/megamek/common/alphaStrike/conversion/ASMovementConverter.java
@@ -163,14 +163,16 @@ final class ASMovementConverter {
     }
 
     private static Map<String, Integer> convertMovementForInfantry(ASConverter.ConversionData conversionData) {
-        ConvInfantry infantry = (ConvInfantry) conversionData.entity();
+        Infantry infantry = (Infantry) conversionData.entity();
         AlphaStrikeElement element = conversionData.element();
         CalculationReport report = conversionData.conversionReport();
 
         var result = new HashMap<String, Integer>();
         int walkingMP = infantry.getWalkMP(MPCalculationSetting.AS_CONVERSION);
         int jumpingMP = infantry.getJumpMP(MPCalculationSetting.AS_CONVERSION);
-        boolean minimalGroundMovement = infantry.hasMinimalGroundMP(MPCalculationSetting.AS_CONVERSION);
+        boolean minimalGroundMovement =
+              infantry instanceof ConvInfantry convInfantry
+                    && convInfantry.hasMinimalGroundMP(MPCalculationSetting.AS_CONVERSION);
         int walkingMove = minimalGroundMovement ? 2 : minimumConvertedMovement(walkingMP * 2);
         int jumpingMove = minimumConvertedMovement(jumpingMP * 2);
