+++ megamek/src/megamek/common/units/ProtoMek.java
@@ -1142,11 +1142,6 @@ public class ProtoMek extends Entity {
     }
 
     @Override
-    public boolean isNuclearHardened() {
-        return true;
-    }
-
-    @Override
     public void setGrappled(int id, boolean attacker) {
         grappled_id = id;
         isGrappleAttacker = attacker;
