+++ megamek/unittests/megamek/common/units/InfantryGliderWingsTest.java
@@ -47,6 +47,7 @@ import megamek.common.options.OptionsConstants;
 import megamek.common.options.PilotOptions;
 import megamek.common.planetaryConditions.Atmosphere;
 import megamek.common.planetaryConditions.PlanetaryConditions;
+import megamek.common.verifier.TestInfantry;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -353,7 +354,7 @@ class InfantryGliderWingsTest {
         void gliderAndFlightWingsMutuallyExclusive() {
             ConvInfantry infantry = createInfantryWithBothWings();
 
-            assertTrue(infantry.hasInvalidWingsConfiguration());
+            assertTrue(TestInfantry.hasInvalidWingsConfiguration(infantry));
         }
 
         @Test
@@ -361,7 +362,7 @@ class InfantryGliderWingsTest {
         void gliderWingsAloneIsValid() {
             ConvInfantry infantry = createInfantry(EntityMovementMode.INF_LEG, true);
 
-            assertFalse(infantry.hasInvalidWingsConfiguration());
+            assertFalse(TestInfantry.hasInvalidWingsConfiguration(infantry));
         }
 
         @Test
